refactor: made edit mode into its separate component
